backend/pyproject.toml
[project]
name = "perception-backend"
version = "1.0.0"
description = "Backend for Perception application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "pydantic[email]>=2.13.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.49",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.44.0",
    "python-multipart>=0.0.26",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
